pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nomad-simulation-examples"
version = "0.1.0"
description = "Example uploads for multiple simulation codes for NOMAD development."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nomad-lab[plugins]>=1.3.0",
]

[project.entry-points."nomad.plugin"]
example_gromacs = "nomad_simulation_examples.example_uploads:gromacs_examples"
example_lammps = "nomad_simulation_examples.example_uploads:lammps_examples"
example_orca = "nomad_simulation_examples.example_uploads:orca_examples"
example_vasp = "nomad_simulation_examples.example_uploads:vasp_examples"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"nomad_simulation_examples" = ["example_uploads/**/*"]
